+++ builder/azure/arm/builder_acc_test.go
@@ -11,7 +11,7 @@ package arm
 //
 // The subscription in question should have a resource group
 // called "packer-acceptance-test" in "South Central US" region. The
-// storage account refered to in the above variable should
+// storage account referred to in the above variable should
 // be inside this resource group and in "South Central US" as well.
 //
 // In addition, the PACKER_ACC variable should also be set to