+++ builder/openstack/builder.go
@@ -126,9 +126,10 @@ func (b *Builder) Run(ctx context.Context, ui packer.Ui, hook packer.Hook) (pack
 			Wait: b.config.RackconnectWait,
 		},
 		&StepAllocateIp{
-			FloatingIPNetwork: b.config.FloatingIPNetwork,
-			FloatingIP:        b.config.FloatingIP,
-			ReuseIPs:          b.config.ReuseIPs,
+			FloatingIPNetwork:     b.config.FloatingIPNetwork,
+			FloatingIP:            b.config.FloatingIP,
+			ReuseIPs:              b.config.ReuseIPs,
+			InstanceFloatingIPNet: b.config.InstanceFloatingIPNet,
 		},
 		&communicator.StepConnect{
 			Config: &b.config.RunConfig.Comm,
@@ -153,6 +154,7 @@ func (b *Builder) Run(ctx context.Context, ui packer.Ui, hook packer.Hook) (pack
 		&stepUpdateImageTags{},
 		&stepUpdateImageVisibility{},
 		&stepAddImageMembers{},
+		&stepUpdateImageMinDisk{},
 	}
 
 	// Run!