+++ builder/qemu/builder_test.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 	"reflect"
 	"testing"
 
@@ -596,5 +597,26 @@ func TestBuilderPrepare_QemuArgs(t *testing.T) {
 
 	if !reflect.DeepEqual(b.config.QemuArgs, expected) {
 		t.Fatalf("bad: %#v", b.config.QemuArgs)
+	}
+}
+
+func TestBuilderPrepare_VNCPassword(t *testing.T) {
+	var b Builder
+	config := testConfig()
+
+	config["vnc_use_password"] = true
+	config["output_directory"] = "not-a-real-directory"
+	b = Builder{}
+	warns, err := b.Prepare(config)
+	if len(warns) > 0 {
+		t.Fatalf("bad: %#v", warns)
+	}
+	if err != nil {
+		t.Fatalf("should not have error: %s", err)
+	}
+
+	expected := filepath.Join("not-a-real-directory", "packer-foo.monitor")
+	if !reflect.DeepEqual(b.config.QMPSocketPath, expected) {
+		t.Fatalf("Bad QMP socket Path: %s", b.config.QMPSocketPath)
 	}
 }